Put the main classes into init
again, this isn't java
