+++ anf/bin/import/xi202_import/xi202_import.xpy
@@ -11,7 +11,7 @@ from optparse import OptionParser
 from pymongo import MongoClient
 import antelope.stock as stock
 from logging import getLogger
-from xi202_import.xi202_import import xi202_importer
+from xi202_import import xi202_importer
 from xi202_import.q330_serials import Q330serials
 from xi202_import.orb_serials import ORBserials
